chore: refactor api/readiness to api/version
